Implement MMU and TLB for Memory Hierarchy Traces (#7)
* chore: remove global import of the parameter.

* feat1: Add TLB into the private memory hierarchy

* feat: refactor data type of block_id

* fix: constant

* feat: add MMU logic

* feat: connect MMU to the design

* feat: Decouple MMU type. The test does not use MMU.

* fix: no copy error

* fix: wrong ttbr + force initialize

* fix: now hint number matches

* feat: force initialization in the beginning

+++ src/components/memory/checkpoint/checkpoint.rs
@@ -30,7 +30,7 @@ pub enum CacheBlockState {
 // The normal cache blocks
 #[derive(PartialEq, Eq, Clone, Debug, Serialize)]
 pub struct CacheBlock {
-    pub block_id: usize,
+    pub block_id: u64,
     pub state: CacheBlockState,
     pub in_instruction_cache: bool,
     pub in_data_cache: bool
@@ -38,7 +38,7 @@ pub struct CacheBlock {
 
 #[derive(PartialEq, Eq, Clone, Debug, Serialize)]
 pub struct DirectoryBlock {
-    pub block_id: usize,
+    pub block_id: u64,
     pub replicas: Vec<u8>, // core_ids
     pub last_writer: Option<u8>
 }
